requirements.txt
# =============================================================================
# Mobile Automation Pipeline — Python Dependencies
# Install with: pip install -r requirements.txt
# =============================================================================

# --- Appium & Selenium -------------------------------------------------------
Appium-Python-Client==3.1.1
selenium==4.18.1

# --- Computer Vision & OCR ---------------------------------------------------
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80         # Headless environments (CI/Replit)
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.4
scikit-image==0.22.0                     # SSIM for screen change detection

# --- HTTP & Async ------------------------------------------------------------
httpx==0.27.0
requests==2.31.0
websockets==12.0
aiohttp==3.9.3

# --- FastAPI & API Server ----------------------------------------------------
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1
websocket-client==1.7.0

# --- Database ----------------------------------------------------------------
SQLAlchemy==2.0.29
alembic==1.13.1
aiosqlite==0.20.0                        # Async SQLite for FastAPI

# --- AWS Device Farm ---------------------------------------------------------
boto3==1.34.69
botocore==1.34.69

# --- Environment & Config ----------------------------------------------------
python-dotenv==1.0.1
click==8.1.7                             # CLI for scripts

# --- Logging & Monitoring ----------------------------------------------------
loguru==0.7.2
rich==13.7.1                             # Beautiful terminal output
structlog==24.1.0

# --- Testing -----------------------------------------------------------------
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
pytest-cov==5.0.0
httpx==0.27.0                            # Test client for FastAPI
responses==0.25.0                        # Mock HTTP responses

# --- Utilities ---------------------------------------------------------------
tenacity==8.2.3                          # Retry logic
python-jose[cryptography]==3.3.0        # JWT for API auth
passlib[bcrypt]==1.7.4
anyio==4.3.0
filelock==3.13.3
pytz==2024.1
